Auto-Generate In-Book ToC (#18)
Auto-Generate the in-book table of contents from the descriptions.json
file. THis will let the Table_of_Contents.md file be linked to the
markdown pages instead of the xhtml pages which will in turn allow me to
update all the chapters to link to the Table_of_Contents.md file instead
of the projects homepage. This will provide mobile readers with a better
experience because clicking on the chapter's home button in the app jumps
the reader to a web-browser instead of staying within the mobile app UI.