Reshi Scene Rewrite
When I read the scene, I felt like Bast should already know what a
Hookah. Additionally, Bast is very old in human years and has "snuck
off" to the human realm multiple times already (e.g. he's left the fey
either unsupervised or with a group of his peers). It seems likely that
he has already encountered a smoke house before and thus would know what
a hookah is. Kvothe, on the otherhand, is young. And though he is
portrayed as world-wise in many things, his character is also portrayed
as niave in other things. Would it make more sense if Kvothe was not
teaching Bast about Hookah's in this scene? Perhaps Kvothe could say
something about always wanting to try a hookah when they pass the
smokehouse? Bast chimes in about how it is a great thing and maybe gives
a little courful context about a time that he and his fey peers snuck
off to the mortal realm to go on a "smoke house" hookah bender and
create mischif? Perhaps he also peer presures Kvothe into going in and
giving it a try? Maybe Bast doesn't know the background of how the Fey
actually gave the Hookah to the humans to help make them less glum, and
so Kvothe is still able to give Bast that background (thus giving Bast a
reason to call Kvothe Reshi and thus preserving the important plot
advancement of the scene).