Reorganize Scenes (#200)
The scene where Kvothe names the wind in Elodin's class should happen
when classes are back in session. Also, move Kvothe's rite of passage
journey to befre classes are back in session which will help to explain
why he has an opertunity to "get away" and also makes it more meaningful
since he is "taming the wind" before he creates his ring of wind in
Elodin's class, becoming officially recognized as a Namer.